tests: move non-string match test to test_nested_filter, cover MatchText and MatchTextAny
